Fix coderunner test
This is possibly a temp fix, there is some weirdness with turning
a simple string in JSON.  JSON.parse doesn't work for whatever is
returned from answer.to_json, so for now removing the quotes was a
solution to get tests passing again.  Also modifies from get to post
based on recent changes.